Move remaining services to own folders
